+++ tests/conftest.py
@@ -1,10 +1,9 @@
 import json
 from pathlib import Path
-from typing import cast
+from typing import Any
 
 import pystac
 import pytest
-from rustac import Item
 
 
 @pytest.fixture
@@ -28,18 +27,18 @@ def data(root: Path) -> Path:
 
 
 @pytest.fixture
-def item(examples: Path) -> Item:
+def item(examples: Path) -> dict[str, Any]:
     with open(examples / "simple-item.json") as f:
         return json.load(f)
 
 
 @pytest.fixture
-def maxar_items(root: Path) -> list[Item]:
+def maxar_items(root: Path) -> list[dict[str, Any]]:
     # https://github.com/stac-utils/rustac/issues/722
     directory = root / "tests" / "data" / "maxar-hurricane-ian-2022"
     item_a = pystac.read_file(directory / "031331303020" / "10300100DB064000.json")
     item_b = pystac.read_file(directory / "031331303211" / "10300100DB064000.json")
     return [
-        cast(Item, item_a.to_dict(transform_hrefs=False)),
-        cast(Item, item_b.to_dict(transform_hrefs=False)),
+        item_a.to_dict(transform_hrefs=False),
+        item_b.to_dict(transform_hrefs=False),
     ]