+++ tests/test_write.py
@@ -1,3 +1,4 @@
+import json
 from pathlib import Path
 from typing import Any
 
@@ -6,7 +7,7 @@ import pyarrow.parquet
 import rustac
 import stac_geoparquet
 from pyarrow.parquet import ParquetFile
-from rustac.store import LocalStore
+from rustac.store import LocalStore, MemoryStore
 
 
 async def test_write(item: dict[str, Any], tmp_path: Path) -> None:
@@ -21,7 +22,7 @@ async def test_write_compressed(item: dict[str, Any], tmp_path: Path) -> None:
     path = str(tmp_path / "out.parquet")
     await rustac.write(path, [item])
     metadata = pyarrow.parquet.read_metadata(path)
-    assert metadata.row_group(0).column(0).compression == "SNAPPY"
+    assert metadata.row_group(0).column(0).compression == "ZSTD"
 
 
 async def test_write_store(item: dict[str, Any], tmp_path: Path) -> None:
@@ -47,3 +48,30 @@ async def test_write_parquet_compression(tmp_path: Path, item: dict[str, Any]) -
     for row_group in range(metadata.num_row_groups):
         for column in range(metadata.num_columns):
             assert metadata.row_group(row_group).column(column).compression == "ZSTD"
+
+
+async def test_write_item_collection_ndjson(
+    tmp_path: Path, item: dict[str, Any]
+) -> None:
+    path = str(tmp_path / "out.ndjson")
+    await rustac.write(path, [item])
+    with open(path) as f:
+        item = json.load(f)
+    assert item["type"] == "Feature"
+
+
+async def test_geoparquet_writer_two(tmp_path: Path, item: dict[str, Any]) -> None:
+    async with rustac.geoparquet_writer([item], str(tmp_path / "out.parquet")) as w:
+        await w.write([item])
+
+    item_collection = await rustac.read(str(tmp_path / "out.parquet"))
+    assert len(item_collection["features"]) == 2
+
+
+async def test_geoparquet_writer_memory_store(item: dict[str, Any]) -> None:
+    store = MemoryStore()
+    async with rustac.geoparquet_writer([item], "out.parquet", store=store) as w:
+        await w.write([item])
+
+    item_collection = await rustac.read("out.parquet", store=store)
+    assert len(item_collection["features"]) == 2